+++ src/routes/(api)/api/v1/episodes/+server.ts
@@ -4,3 +4,5 @@ import { ProfessionRepository } from '$lib/data/repository/ProfessionRepository'
 export function GET() {
 	return json(ProfessionRepository.professions);
 }
+
+export const prerender = true;
