+++ src/routes/(api)/api/v1/items/+server.ts
@@ -1,0 +1,8 @@
+import { json } from '@sveltejs/kit';
+import { ItemRepository } from '$lib/data/repository/ItemRepository';
+
+export function GET() {
+	return json(ItemRepository.items);
+}
+
+export const prerender = true;
