+++ src/routes/(api)/api/v1/professions/+server.ts
@@ -1,0 +1,8 @@
+import { json } from '@sveltejs/kit';
+import { ProfessionRepository } from '$lib/data/repository/ProfessionRepository';
+
+export function GET() {
+	return json(ProfessionRepository.professions);
+}
+
+export const prerender = true;
