--- src/routes/api/+layout.ts
@@ -1,1 +1,0 @@
-export const ssr = false;
