--- src/routes/api/+page.svelte
@@ -1,26 +1,0 @@
-<script lang="ts">
-	import { base } from '$app/paths';
-</script>
-
-<h2>Overviews</h2>
-<ul>
-	<li><a href="{base}/api/items">Items</a></li>
-	<li><a href="{base}/api/episodes">Episodes</a></li>
-	<li><a href="{base}/api/professions">Professions</a></li>
-</ul>
-
-<h2>Examples</h2>
-<ul>
-	<li><a href="{base}/api/item/bread">{base}/api/item/bread</a></li>
-	<li><a href="{base}/api/episode/hopeport">{base}/api/episode/hopeport</a></li>
-	<li><a href="{base}/api/profession/fisher">{base}/api/profession/fisher</a></li>
-</ul>
-
-<p>
-	This links don't appear to work in the browser because I am a SvelteKit novice but they work on GET requests I
-	promise.
-</p>
-<p>
-	Stop complaining and
-	<a href="https://github.com/Ishadijcks/BrighterShores/issues/7" target="_blank">help me fix it!</a>
-</p>
