--- src/routes/api/[...profession=profession]/+server.ts
@@ -1,7 +1,0 @@
-import { json } from '@sveltejs/kit';
-import { ProfessionRepository } from '$lib/data/repository/ProfessionRepository';
-import type { ProfessionId } from '$lib/data/game/ProfessionId';
-
-export function GET({ params }) {
-	return json(ProfessionRepository.getProfession(('/' + params.profession) as ProfessionId));
-}
