--- src/routes/api/episodes/+server.ts
@@ -1,6 +1,0 @@
-import { json } from '@sveltejs/kit';
-import { EpisodeRepository } from '$lib/data/repository/EpisodeRepository';
-
-export function GET() {
-	return json(EpisodeRepository.episodes);
-}
