--- src/routes/api/items/+server.ts
@@ -1,6 +1,0 @@
-import { ItemRepository } from '$lib/data/repository/ItemRepository';
-import { json } from '@sveltejs/kit';
-
-export function GET() {
-	return json(ItemRepository.items);
-}
